second_device

+++ 6.4_PWM控制舵机/Hardware/Servo.c
@@ -12,6 +12,6 @@ void Servo_SetAngle(int8_t angle)
     // 舵机的PWM周期为20ms
     // 角度值范围为-90到+90度
     // 对应的PWM脉宽范围为0.5ms到2.5ms
-    int16_t compareValue = 1500 + angle* 2000 / 180 ; // 1ms + (angle/180)*1ms
+    int8_t compareValue = 1500 + angle* 2000 / 180 ; // 1ms + (angle/180)*1ms
     PWM_SetCompare2(compareValue);
 }
